+++ test/src/main/java/org/linguafranca/pwdb/checks/DatabaseLoaderChecks.java
@@ -23,6 +23,7 @@ import org.linguafranca.pwdb.*;
 import java.io.PrintStream;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
+import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
 
@@ -88,6 +89,63 @@ public abstract class DatabaseLoaderChecks <D extends Database<D,G,E,I>, G exten
 
         Assert.assertEquals(1, entries.size());
         assertEquals("pass", entries.get(0).getPassword());
+
+        // kdb files don't have a time zone so can't make head or tail of the date - test file seems to have a local time in it
+        if (skipDateCheck) {
+            return;
+        }
+
+        Date c = entries.get(0).getCreationTime();
+        Date expected = sdf.parse("2015-10-24T17:20:41Z");
+        Assert.assertEquals(expected, c);
+    }
+
+    @Test
+    public void test123FileBytePassword() throws ParseException {
+
+        // visit all groups and entries and list them to console
+        database.visit(new Visitor.Print(printStream));
+
+        // find all entries in the database
+        // the kdb version has three additional system related entries
+        List<? extends E> anything = database.findEntries("");
+        Assert.assertTrue(10 <= anything.size());
+
+        // find all entries in the database that have the string "test" in them
+        List<? extends E> tests = database.findEntries("test");
+        for (Entry tes: tests) {
+            printStream.println(tes.getTitle());
+        }
+        Assert.assertEquals(4, tests.size());
+        if (tests.size() > 0) {
+            // copy the password of the first entry to the clipboard
+            byte[] pass = tests.get(0).getPasswordAsBytes();
+/*
+            StringSelection selection = new StringSelection(pass);
+            Toolkit.getDefaultToolkit().getSystemClipboard().setContents(selection, selection);
+            printStream.println(pass + " copied to clip board");
+*/
+            // all the relevant entries should have the password 123
+            byte[] pass2 = tests.get(0).getPasswordAsBytes();
+            Assert.assertEquals(true, Arrays.equals(pass, pass2));
+            Assert.assertEquals("123", new String(pass2));
+        }
+
+        List<? extends E> passwords = database.findEntries("password");
+        Assert.assertEquals(4, passwords.size());
+        for (Entry passwordEntry : passwords) {
+            Assert.assertEquals(true, Arrays.equals(passwordEntry.getTitle().getBytes(), passwordEntry.getPasswordAsBytes()));
+            printStream.println(passwordEntry.getTitle());
+        }
+
+        List<? extends E> entries = database.findEntries(new Entry.Matcher() {
+            @Override
+            public boolean matches(Entry entry) {
+                return entry.getTitle().equals("hello world");
+            }});
+
+        Assert.assertEquals(1, entries.size());
+        assertEquals("pass", new String(entries.get(0).getPasswordAsBytes()));
 
         // kdb files don't have a time zone so can't make head or tail of the date - test file seems to have a local time in it
         if (skipDateCheck) {